detector/requirements.txt
pyyaml
requests
flask
psutil
gunicorn
